dev: beautify charts and code style
